Move `getVTimeZoneId` from `Collection` to `ICalendarUtils`
